chore(homepage): remove sound add dynamic imports update readme update pic whiteboard
