+++ src/components/RenderModel.jsx
@@ -1,8 +1,6 @@
 'use client';
-// import { Environment } from '@react-three/drei';
 import { Canvas } from '@react-three/fiber';
 import React, { Suspense } from 'react';
-// import { Html } from '@react-three/drei';
 
 const RenderModel = ({ children }) => {
 
@@ -18,7 +16,7 @@ const RenderModel = ({ children }) => {
             <Suspense fallback={null}>
                 {children}
             </Suspense>
-            {/* <Environment preset="sunset" /> */}
+
             <ambientLight intensity={0.5} />
             <directionalLight position={[3, 2, 5]} intensity={1} />
         </Canvas>
